add: dialog examples
